requirements.txt
# Azure SDKs
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
openai>=1.40.0

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Image Processing
opencv-python-headless>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0

# Data Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
pyodbc>=5.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.27.0

# Logging & Monitoring
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
